fix: split one common reducer to different reducers
